requirements.txt
Django==2.2
djangorestframework==3.9.2
django-rest-swagger==2.2.0
djangorestframework==3.9.2
drf_yasg==1.15.0